Widen EC completions to u16 + EC1 goal-1000 inside Enslaved (4.5/7.3)
The original's maxCompletions is dynamic: Enslaved.isRunning && id === 1 ?
1000 : 5. Add ec_max_completions(id) and thread it through the goal formula
(goalAtCompletions clamps at max-1, so EC1's goal scales to 1e1800 ×
(1e200)^999 inside the run), completion banking, the pending-completion scan,
and EC1's study secondary requirement (min(completions, 999) × 20000).

Widen the completion storage u8 → u16 across GameState, the glyph-undo
snapshot, the automator's last_ec_completions, the GUI view, and the save
decode (which clamps EC1 at 1000, others at 5).

Fidelity suite unchanged at 1469/1476.

+++ crates/ad-core/src/automator/mod.rs
@@ -211,7 +211,7 @@ pub struct AutomatorRuntime {
     /// `findLastPrestigeRecord` log text.
     pub last_prestige_gain: [Decimal; 3],
     /// EC completions banked by the last Eternity (`lastECCompletionCount`).
-    pub last_ec_completions: u8,
+    pub last_ec_completions: u16,
 }
 
 impl AutomatorData {
